fix: move twig "snippet" function from "twig_plus" to "twig_core" plugin
