+++ tests/acceptance/HerbieInfoCest.php
@@ -428,6 +428,7 @@ final class HerbieInfoCest
             'page_link',
             'output_css',
             'output_js',
+            'snippet',
             'translate',
             'url',
             'abs_url',
@@ -442,7 +443,6 @@ final class HerbieInfoCest
             'pages_recent',
             'page_title',
             'sitemap',
-            'snippet',
             'taxonomy_archive',
             'taxonomy_authors',
             'taxonomy_categories',
